Implement snippet saving (#2)
Initially, the website currently at rouge.jneen.net supported the saving
of snippets. This was achieved via an SQL database. Gradually, the
number of snippets saved exceeded the maximum number of rows in
Heroku's free tier. After that point, the saving feature no longer
worked.

This commit implements snippet saving but by another method. Rather than
saving the snippet to a database, the text of the snippet is base-64
encoded and provided as part of a URL. While this does limit the maximum
length of a snippet (most systems support a maximum URL length of around
2000 characters), it is not considered a problem because the new dingus
already limits the maximum length of the text parsed for performance
reasons.

This commit implements both client-side and server-side encoding of the
URL component. If JavaScript is enabled, clicking the 'Save this' button
will change the browser window's URL and save the encoded URL to the
clipboard. If JavaScript is not enabled, the source text will be sent to
the server where the URL component is encoded and the user redirected to
a URL that includes the encoded result.