Support old pastes (#6)
The website currently at rouge.jneen.net previously allowed users to                 
save snippets of source code that could then be retrieved via a unique               
URL. Although users can no longer save snippets, the existing URLs still             
work and continue to be used (primarily by the maintainers of Rouge as
they were often included in bug reports filed on GitHub). If the dingus              
is to fully replace the current website, it should display these                     
snippets.
The current website uses a PostgreSQL database to store the data. The
dingus instead uses the GDBM library. This keeps the system light and
simple and makes development easier.
In addition:
1. Although no guarantees of privacy were ever given, users who saved                
   snippets may not have appreciated that they could be put in a 
   publicly accessible Git repository one day. To provide some
   protection, the dingus stores an encrypted version of the GDBM
   database file. The data is encrypted using the Lockbox gem.
2. A decrypted version of the database is created on first access and
   saved to the file system. No checks have been done as to the thread               
   safety of this design.
3. To fully use the system locally now requires two environmental
   variables: (1) a `HASHIDS_SALT` string for use in decoding the URL
   slugs; and (2) a `LOCKBOX_KEY` string for use in the decryption of
   the GDBM database file. In development, these can be saved in the
   root-level file `.env` and will be pulled in by the dotenv gem.
4. The design of the paste page differs slightly from that used on the
   current website.
It is noted that URLs for old snippets could redirect to a URL where the
code in the snippet is encoded in the URL itself (as is the case with
new snippets). Since old snippets were not subject to the same length
requirements and this could be buggy, this has been left as a possible
future improvement.