Improve editing experience (#7)
This PR improves the JavaScript code used in the dingus in four ways.

First, the dingus will no longer overwrite the content of the `textarea`
element as a user is typing. In previous versions, the value of the
`textarea` would be overwritten whenever content was received from the
server. As a user typed, the JavaScript in the client would be sending
and receiving highlighted text and so there was always a risk that what
was received from the server was now 'behind' what the user had typed.
Indeed, given that the dingus rate limits what it sends to the server, a
fast typist would almost certainly see characters 'disappear'. The
dingus now leaves the content of the `textarea` alone.

Second, the dingus will no longer overwrite the content of the
`textarea` element if the user has typed text. In previous versions,
swapping the highlighting language would cause the value of the
`textarea` to be replaced by the demo text for that language. That now
only happens if the text in the `textarea` has not been edited or is
blank.

Third, error handling has been improved. There are now centralised
settings for the maximum URL path and body content sizes that are
referred to from both the Ruby and JavaScript code. Error messages are
similarly centralised. Finally, there is visual feedback provided by
disabling the save button when the length of the user's text exceeds the
maximum length.

Fourth, the highlighted code now scrolls in tandem with the `textarea`.
This works both as the user types new code but also if the user scrolls
the `textarea`.